requirements.txt
llama-index 
llama-index-vector-stores-qdrant 
fastembed 
llama-index-embeddings-fastembed 
llama-index-llms-groq 
llama-index-vector-stores-chroma
python-dotenv
streamlit
chroma
pysqlite3-binary
llama-index-embeddings-huggingface
sentence-transformers